pyproject.toml
[project]
name = "colorblind-snapshot-analyzer"
version = "0.1.0"
description = ""
authors = [
    {name = "NotTheRealWallyx",email = "nottherealwallyx@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.2.1,<12.0.0)",
    "daltonize (>=0.2.0,<0.3.0)",
    "requests (>=2.32.4,<3.0.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "setuptools (>=80.9.0,<81.0.0)"
]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pre-commit = "^4.2.0"
isort = "^6.0.1"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

packages = [{ include = "colorblind_snapshot_analyzer" }]